Part 2 completed
Here we have created a deployment which containerizes the application from Part 1.
Instead of triggering the application via the host machine, we push the application by creating a container image to the K8s cluster.
This images was push to the image repository docker hub.
There were the code adjustment which made it possible to get the config via the rest.InClusterConfig() method along with the roles and rolebinding necessary to authenticate the serviceaccount.
One done, we were able to list the pods and deployment within the k8s cluster as well.