Part3 completed
Part3 was mainly about unstanding how a K8s object/resource is described in the Go code.
If a Go struct implements the runtime.Object interface from apimachinery package, this means that the struct as a kubernetes object.
Detailed explaination for the above:
If a Go struct implements the SetGroupVersionKind(), GroupVersionKind() and DeepCopyObject() methods then that struct can be considered as a kubernetes object/resource.
We took a look at the Pod struct in Go as an example to understand a k8s resource/object